+++ src/main/java/com/example/quaterback/login/repository/refresh/SpringDataRefreshRepository.java
@@ -3,7 +3,7 @@ package com.example.quaterback.login.repository.refresh;
 import com.example.quaterback.login.entity.RefreshEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RefreshRepository extends JpaRepository<RefreshEntity, Long> {
+public interface SpringDataRefreshRepository extends JpaRepository<RefreshEntity, Long> {
 
     Boolean existsByRefresh(String refresh);
